requirements.txt
flask
datetime
WTForms
Flask_WTF
python-dotenv
email_validator
flask_sqlalchemy
flask_bcrypt
Flask-Login